ARM: dts: msm: enable SDR104 workaround for MDM9x55
On MDM9x55, its observed that the first SDIO command after system
suspend-resume is failing with a timeout. This issue is seen only
with SDR104 mode.

In the SDIO suspend path, the gcc_sdcc1_apps clock would be gated.
The DLL which used in SDR104 getting freeze when clocks are gated
is resulting in command timeouts in SDR104 mode after system resume.

To mitigate this issue, s/w has to retune the DLL every time clocks
are gated dynamically in SDR104 mode.

This SDR104_wa flag enables needed logic to tune the DLL during
clock un-gating path.

Change-Id: I77dc52d31fe989e192ac45b2cd19e5d9faf463fa
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>